Move cli to a separate library
