+++ build.sbt
@@ -127,6 +127,7 @@ lazy val enso = (project in file("."))
     graph,
     logger.jvm,
     pkg,
+    cli,
     `version-output`,
     runner,
     runtime,
@@ -475,6 +476,18 @@ lazy val pkg = (project in file("lib/scala/pkg"))
     libraryDependencies ++= circe ++ Seq(
         "io.circe"  %% "circe-yaml" % circeYamlVersion, // separate from other circe deps because its independent project with its own versioning
         "commons-io" % "commons-io" % commonsIoVersion
+      )
+  )
+  .settings(licenseSettings)
+
+lazy val cli = project
+  .in(file("lib/scala/cli"))
+  .configs(Test)
+  .settings(
+    version := "0.1",
+    libraryDependencies ++= Seq(
+        "org.scalatest" %% "scalatest" % scalatestVersion % Test,
+        "org.typelevel" %% "cats-core" % catsVersion
       )
   )
   .settings(licenseSettings)
@@ -980,5 +993,6 @@ lazy val launcher = project
         .value
   )
   .settings(licenseSettings)
+  .dependsOn(cli)
   .dependsOn(`version-output`)
   .dependsOn(pkg)
