--- engine/launcher/src/main/scala/org/enso/launcher/cli/internal/AdditionalArguments.scala
@@ -1,14 +1,0 @@
-package org.enso.launcher.cli.internal
-
-class AdditionalArguments(helpComment: String) extends BaseOpts[Seq[String]] {
-  var value: Seq[String]                        = Seq()
-  override private[cli] val additionalArguments = Some(args => value = args)
-
-  override private[cli] def result()      = Right(value)
-  override private[cli] def reset(): Unit = value = Seq()
-
-  override def availableOptionsHelp(): Seq[String] =
-    if (helpComment.nonEmpty) {
-      Seq(helpComment)
-    } else Seq()
-}
