--- engine/launcher/src/main/scala/org/enso/launcher/cli/internal/OptsPure.scala
@@ -1,8 +1,0 @@
-package org.enso.launcher.cli.internal
-
-class OptsPure[A](v: A) extends BaseOpts[A] {
-  override private[cli] def result() = Right(v)
-  override private[cli] def reset(): Unit = {}
-
-  override def availableOptionsHelp(): Seq[String] = Seq()
-}
